Create abstract BasePropertiesModel model
