+++ nautobot/extras/models/roles.py
@@ -1,15 +1,9 @@
 from django.contrib.contenttypes.models import ContentType
 from django.db import models
 
-from nautobot.core.fields import AutoSlugField
-from nautobot.core.models import BaseModel
 from nautobot.extras.fields import LimitedChoiceField
-from nautobot.extras.models import ChangeLoggedModel, CustomFieldModel, RelationshipModel
-from nautobot.extras.models.mixins import NotesMixin
-from nautobot.extras.querysets import ContentTypeRelatedQuerySet
+from nautobot.extras.models import BasePropertiesModel
 from nautobot.extras.utils import RoleModelsQuery, extras_features
-from nautobot.utilities.choices import ColorChoices
-from nautobot.utilities.fields import ColorField
 
 
 @extras_features(
@@ -21,10 +15,7 @@ from nautobot.utilities.fields import ColorField
     "relationships",
     "webhooks",
 )
-class Role(BaseModel, ChangeLoggedModel, CustomFieldModel, RelationshipModel, NotesMixin):
-    # TODO(timizuo): Create abstract model for this because Role, Status and Tags
-    #  all contain similar fields
-
+class Role(BasePropertiesModel):
     content_types = models.ManyToManyField(
         to=ContentType,
         related_name="roles",
@@ -32,41 +23,6 @@ class Role(BaseModel, ChangeLoggedModel, CustomFieldModel, RelationshipModel, No
         limit_choices_to=RoleModelsQuery(),
         help_text="The content type(s) to which this role applies.",
     )
-    name = models.CharField(max_length=50, unique=True)
-    color = ColorField(default=ColorChoices.COLOR_GREY)
-    slug = AutoSlugField(populate_from="name", max_length=50)
-    description = models.CharField(
-        max_length=200,
-        blank=True,
-    )
-
-    objects = ContentTypeRelatedQuerySet.as_manager()
-
-    csv_headers = ["name", "slug", "color", "content_types", "description"]
-    clone_fields = ["color", "content_types"]
-
-    class Meta:
-        ordering = ["name"]
-
-    def __str__(self):
-        return self.name
-
-    def natural_key(self):
-        return (self.name,)
-
-    # TODO(timizuo): When view url has been implemented for role; visit this
-    # def get_absolute_url(self):
-    #     return reverse("extras:status", args=[self.slug])
-
-    def to_csv(self):
-        labels = ",".join(f"{ct.app_label}.{ct.model}" for ct in self.content_types.all())
-        return (
-            self.name,
-            self.slug,
-            self.color,
-            f'"{labels}"',  # Wrap labels in double quotes for CSV
-            self.description,
-        )
 
 
 class RoleField(LimitedChoiceField):
